+++ library/src/device/solution-shipping.py
@@ -297,7 +297,8 @@ def cli():
     if 'all' in archs:
         archs = [
             'gfx900', 'gfx906', 'gfx908', 'gfx90a', 'gfx940', 'gfx941',
-            'gfx942', 'gfx1030', 'gfx1100', 'gfx1101', 'gfx1102'
+            'gfx942', 'gfx1030', 'gfx1100', 'gfx1101', 'gfx1102', 'gfx1200',
+            'gfx1201'
         ]
 
     # remove xnack and sramecc