+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Documentation for rocFFT is available at
 ### Changes
 
 * Add --smoketest option to rocfft-test.
+* Support gfx1200 and gfx1201 architectures.
 
 ## rocFFT 1.0.28 for ROCm 6.2.0